added tailwind possibly working properly
