--- Frontend/src/About.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-class About extends Component {
-  render() {
-    return (
-        <div>
-            <h1 className='text-xl'>Hello this is about page.</h1>
-        </div>
-    )
-  }
-}
-
-export default About;
