--- Frontend/src/Home.js
@@ -1,11 +1,0 @@
-
-
-function Home() {
-  return (
-    <div class="bg-slate-600 w-14 h-14">
-      <h1>This is the home page</h1>
-    </div>
-  );
-}
-
-export default Home;
