+++ Frontend/src/MainArea/Home.js
@@ -1,0 +1,9 @@
+function Home() {
+  return (
+    <div className="bg-slate-600 w-full h-full">
+      <h1>This is the home page</h1>
+    </div>
+  );
+}
+
+export default Home;
