+++ Frontend/src/Navbar/NavBar.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import { Link } from "react-router-dom";
+
+class Navbar extends Component {
+  render() {
+    return (
+        <div className='bg-slate-600 w-full h-12'>
+             <h1>Hello world!</h1>
+        </div>
+    )
+  }
+}
+
+export default Navbar;
