+++ Frontend/src/Sidebar/Sidebar.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+
+class Sidebar extends Component {
+  render() {
+    return (
+        <div className='bg-slate-300 w-full h-8'>
+             <h1>Sidebar</h1>
+            
+        </div>
+    )
+  }
+}
+
+export default Sidebar;
