+++ Frontend/tailwind.config.js
@@ -2,6 +2,10 @@
 module.exports = {
   content: [
     "./src/**/*.{js,jsx,ts,tsx}",
+    "./src/MainArea/*.{js,jsx,ts,tsx}",
+    "./src/Navbar/*.{js,jsx,ts,tsx}",
+    "./src/Sidebar/*.{js,jsx,ts,tsx}",
+    './public/*.html',
   ],
   theme: {
     extend: {},
